refactoring: screen generator integration with ini files
