+++ java_console/.idea/runConfigurations/IniFileModel.xml
@@ -1,22 +1,16 @@
 <component name="ProjectRunConfigurationManager">
   <configuration default="false" name="IniFileModel" type="Application" factoryName="Application" nameIsGenerated="true">
-    <extension name="coverage" enabled="false" merge="false" sample_coverage="true" runner="idea">
+    <option name="MAIN_CLASS_NAME" value="com.opensr5.ini.IniFileModel" />
+    <module name="ui" />
+    <option name="WORKING_DIRECTORY" value="file://$PROJECT_DIR$" />
+    <extension name="coverage">
       <pattern>
-        <option name="PATTERN" value="com.rusefi.ui.config.*" />
+        <option name="PATTERN" value="com.opensr5.ini.*" />
         <option name="ENABLED" value="true" />
       </pattern>
     </extension>
-    <option name="MAIN_CLASS_NAME" value="com.rusefi.ui.config.IniFileModel" />
-    <option name="VM_PARAMETERS" value="" />
-    <option name="PROGRAM_PARAMETERS" value="" />
-    <option name="WORKING_DIRECTORY" value="file://$PROJECT_DIR$" />
-    <option name="ALTERNATIVE_JRE_PATH_ENABLED" value="false" />
-    <option name="ALTERNATIVE_JRE_PATH" value="" />
-    <option name="ENABLE_SWING_INSPECTOR" value="false" />
-    <option name="ENV_VARIABLES" />
-    <option name="PASS_PARENT_ENVS" value="true" />
-    <module name="ui" />
-    <envs />
-    <method />
+    <method v="2">
+      <option name="Make" enabled="true" />
+    </method>
   </configuration>
 </component>
