+++ java_console/inifile/src/com/opensr5/ini/DialogModel.java
@@ -1,4 +1,4 @@
-package com.rusefi.ui.config;
+package com.opensr5.ini;
 
 import java.util.ArrayList;
 import java.util.List;
