+++ java_console/opensr5/src/com/opensr5/Main.java
@@ -1,8 +1,8 @@
 package com.opensr5;
 
-import com.opensr5.io.IniFileMetaInfo;
-import com.opensr5.io.IniFileReader;
-import com.opensr5.io.RawIniFile;
+import com.opensr5.ini.IniFileMetaInfo;
+import com.opensr5.ini.IniFileReader;
+import com.opensr5.ini.RawIniFile;
 
 import java.io.File;
