+++ java_console/ui/src/com/rusefi/ui/SettingsTab.java
@@ -1,5 +1,8 @@
 package com.rusefi.ui;
 
+import com.opensr5.ini.DialogModel;
+import com.opensr5.ini.IniFileModel;
+import com.rusefi.Launcher;
 import com.rusefi.config.Field;
 import com.rusefi.config.FieldType;
 import com.rusefi.config.FieldsMap;
@@ -15,8 +18,8 @@ import java.util.ArrayList;
 import java.util.Map;
 import java.util.concurrent.atomic.AtomicBoolean;
 
-import static com.rusefi.ui.config.IniFileModel.RUSEFI_INI_PREFIX;
-import static com.rusefi.ui.config.IniFileModel.RUSEFI_INI_SUFFIX;
+import static com.opensr5.ini.IniFileModel.RUSEFI_INI_PREFIX;
+import static com.opensr5.ini.IniFileModel.RUSEFI_INI_SUFFIX;
 
 /**
  * @see EnumConfigField
@@ -36,7 +39,7 @@ public class SettingsTab {
     }
 
     public void showContent() {
-        final Map<String, DialogModel> dialogs = IniFileModel.getInstance().getDialogs();
+        final Map<String, DialogModel> dialogs = IniFileModel.getInstance(Launcher.INI_FILE_PATH).getDialogs();
         if (dialogs.isEmpty()) {
             content.removeAll();
             content.add(new JLabel("Meta data not found: " + RUSEFI_INI_PREFIX + "*" + RUSEFI_INI_SUFFIX));
@@ -128,7 +131,7 @@ public class SettingsTab {
         dialog.setText(name);
         dialogBody.removeAll();
 
-        DialogModel m = IniFileModel.getInstance().getDialogs().get(name);
+        DialogModel m = IniFileModel.getInstance(Launcher.INI_FILE_PATH).getDialogs().get(name);
 
         dialogBody.setLayout(new GridLayout(m.getFields().size(), 1));
