+++ java_console/ui/ui.iml
@@ -21,5 +21,6 @@
     <orderEntry type="module" module-name="romraider" />
     <orderEntry type="library" name="jlatexmath" level="project" />
     <orderEntry type="module" module-name="configuration_definition" />
+    <orderEntry type="module" module-name="inifile" />
   </component>
 </module>
